Move AxesGeometrySource to core geometric sources
